Began writing the parser.
After researching the standard way of writing a compiler I have decided to build a parser with the goal of generating an AST. I have also moved the code responsible for printing the test outputs to the Program.cs script. I have also removed A LOT of code and comments; ideas that I chose not to act on.